% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-SBmedian.R
\docType{package}
\name{SBmedian}
\alias{SBmedian}
\alias{SBmedian-package}
\title{Scalable Bayes with Median of Subset Posteriors}
\description{
Median-of-means is a generic yet powerful framework for scalable and robust estimation. 
A framework for Bayesian analysis is called M-posterior, which estimates a median of subset posterior measures.
}
